Refactor: refactored UserForm component into sub components
